+++ stormify/src/jvmBasedMain/kotlin/onl/ycode/stormify/coroutines/ConnectionElement.jvmBased.kt
@@ -16,6 +16,7 @@ import kotlin.coroutines.CoroutineContext
 internal actual class ConnectionElement actual constructor(
     actual internal val conn: Connection,
     actual internal val stormify: Stormify,
+    actual internal val inTransaction: Boolean,
 ) : ThreadContextElement<Unit>, CoroutineContext.Element {
 
     actual override val key: CoroutineContext.Key<*> get() = Key